requirements.txt
anthropic>=0.39.0
gradio>=4.0.0
pandas>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
cryptography>=41.0.0
requests>=2.31.0
websockets>=12.0
pyyaml>=6.0
openai>=1.0.0
google-generativeai>=0.3.0
boto3>=1.28.0
psycopg2-binary>=2.9.0
redis>=5.0.0
